Bare minimum kernel and bootable image
This commit creates a custom target defined in
"x64_64-rust_os.json" which is suitable for compiling
our kernel against.

Notably, it disables SIMD features (as the context switching
on interrupt is expensive when using SIMD registers), enables
using non-SIMD registers for floating point operations, disables
temporary usage of memory beyond the stack pointer (as it could
result in corruptions when it is overwritten by interrupt routines)
and it uses the cross-platform LLD linker.

The bare-minimum freestanding binary now writes
"Hello world!" into the VGA buffer.

When "cargo run" is called, the bare minimum kernel is compiled,
the "bootimage" tool links the resulting executable to the
bootloader dependency to create a bootable image which is
run in QEMU.