Extracted VGA print functionality to new module
The unsafe operations required to write to a raw pointer
are restricted to this module, meaning users of the module
do not have to use unsafe blocks.

A static writer is defined, made public and protected by
a spinlock.

print! and println! macros which call this module's functionality
are made available to the rest of the crate.