Test framework in a bare metal environment
The custom_test_framework feature is used to create our
own test framework which does not depend on the standard library,
as the default test framework does depend on the standard library.

The isa-debug-exit device is used by the kernel to exit QEMU
immediately after the tests are complete, so the test does not loop
forever.

The test functionality is mainly defined in lib.rs which is used
by the unit tests in main.rs and the integration tests in the tests
directory