Implemented breakpoint exception handler
This is the first x86-64 exception our kernel handles.

We create an interrupt descriptor table (IDT) and set
its breakpoint exception handler to a function which
prints a message to the VGA buffer. The IDT is then loaded
on the CPU.

We test this works by invoking a breakpoint (int3)
instruction, then checking that the kernel continues
(which it will do if the breakpoint handler runs successfully).