Experiment with unhandled faults
This commit triggers a page fault, which in turn causes a double
fault and triple fault as the IDT does not have handlers for these
faults.

This causes QEMU to continuously restart on 'cargo run' as like most
hardware, a triple fault causes a restart.

Note that this commit causes the tests to fail (for educational purposes)
but they will be fixed later.