Run double fault handlers on known good stack
In order to allow the double fault handler to succeed in cases
where the existing stack is in a bad state (like a stack overflow)
it needs to operate on a different, known good stack.

This is done by creating a new known good stack in an interrupt
stack table, creating the supporting TSS and GDT data structures,
loading them onto the CPU and changing the double fault's IDT entry
to refer to such a stack in the interrupt stack table.

Also removes stack overflow from main function, as that experiment
is finished and I want the tests to pass again.