Added timer interrupt handler
This prints a dot to the VGA buffer on every timer interrupt. It
also sends the end of interrupt signal to the PIC so that it knows
the interrupt has been handled, and that it can send more. If the EOI
signal was not sent, only one dot would be printed.