Added keyboard interrupt handler
The handler function interacts with the PS/2 keyboard controller.
The motherboard provides a PS/2 emulation layer for USB keyboards
so this handler works for USB keyboards as well.

The handler prints the digits 0-9 when they are pressed by reading
the scancode from the PS/2 controller's data port, then replying with
the EOI signal.