Added dummy page fault handler
Investigate how page faults work by creating a dummy page fault
handler which prints the address which caused the page fault (this
is obtained by reading the CR2 register) and the operation which
caused the page fault (this is described in the page fault error
code which is passed to the page fault handler).

Also trigger a page fault by writing to an unmapped address, to view
the output of this page fault handler.