Map physical memory to virtual address space
This is done by enabling the "map_physical_memory" feature
of the bootloader crate.

Additionally, this commit uses the entry_point! macro provided
by the bootloader crate which type checks the entry point function.
This eliminates the need for explicitly specifying the C calling convention,
calling the entry point '_start', and specifying the name should not
be mangled, as the macro generates this code automatically.