Implement functionality to query page tables
The physical address of the active level 4 page table is stored
in the CR3 register. In order to access the level 4 page table,
it must be accessed with a virtual address. Since the entirety of
physical memory is mapped into virtual memory, the virtual addresses
of the page tables can be calculated from the physical addresses.

This commit accesses the active level 4 page table, and the active
level 3 page tables, and prints out the used entries.