Create trivial example page table mapping
This is done by using the map_to method of the Mapper trait
to map the first page in the virtual address space, to the frame
of the VGA text buffer.

To test that this mapping was successful, the first page in the virtual
address space is written to, and we should see the output in the VGA text
buffer.

The map_to method requires a frame allocator (as new page tables may
need to be created), however since the bootloader loads itself into the
first page of the virtual address space, this page already has all the
required page table entries in physical memory. Therefore we do not need
a functional frame allocator, so we simply define a dummy one which always
returns None.

Note that the map_to method is not safe, as the caller must guarantee that
the frame being mapped to is not currently mapped.