Implement proper frame allocator
The BootInfoFrameAllocator takes the memory map provided by
the BIOS firmware, and uses it to create an iterator over the
addresses of unused physical frames.

Since this frame allocator works properly, the map_to method
does not encounter a frame allocation failure, and we see the
'New!' text in the VGA text buffer as expected.