Add basic bump allocator
This allocates memory at the 'next' pointer on alloc. A severe
drawback is that it cannot free memory until all active allocations
have been freed. This is higlighted by the addition of a test which
performs a single long-standing allocation, then executes a loop
which allocates/frees a Box. This will eventually cause an OOM error
as the long-standing allocation will not allow the bump allocator to
free any memory.