Bare bones cooperative multitasking API
This commit provides the Task API which is a thin
wrapper around Rust's async/await mechanisms.

A Task takes a Future with a static lifetime and
no return value, and provides a poll method
which invokes the poll method of the passed
Future to enable an executor to check if
the Future has completed yet.

The SimpleExecutor maintains a queue of Tasks
and continuously polls them, readding them to
the back of the queue if they have not completed.