Implement asynchronous keyboard input handling with Waker notifications
The keyboard interrupt handler will place the keypresses on a queue
which does not block or allocate (to avoid deadlocking).

The executor will run the print_keypresses Task, which takes keypresses
from the ScancodeStream (which in turn takes them from the queue).

The ScancodeStream and keyboard interrupt handler use Waker notifications
so the executor does not have to poll the Task, and instead it is notified
when there is a key to be processed.