Create more efficient Executor which uses Waker notifications instead of busy-waiting
The Executor runs Tasks whose TaskIds are on its task_queue, and a pending Task gets woken
by an interrupt handler by adding its TaskId to this queue.

Instead of continuously polling the task_queue, the Executor executes the hlt instruction
which lets the CPU sleep until the next interrupt. Since TaskIds are added to the task_queue
by interrupt handlers, this approach is much more CPU-efficient, and you do not sacrifice
responsiveness.